requirements.txt
Django==5.0.3
PyJWT==2.9.0
asgiref==3.8.1
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
gunicorn==21.2.0
packaging==24.0
channels==3.0.5
